requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.1.1
protobuf==3.18.1
python-telegram-bot==14.2
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
